Move zeroSync logic out, refactor assignAccountFromPrivateKey
