+++ packages/mobile/src/redux/sagas.ts
@@ -19,7 +19,7 @@ import { networkInfoSaga } from 'src/networkInfo/saga'
 import { sendSaga } from 'src/send/saga'
 import { stableTokenSaga } from 'src/stableToken/saga'
 import Logger from 'src/utils/Logger'
-import { syncModeSaga } from 'src/web3/saga'
+import { web3Saga } from 'src/web3/saga'
 
 const loggerBlacklist = [
   'persist/REHYDRATE',
@@ -73,5 +73,5 @@ export function* rootSaga() {
   yield spawn(dappKitSaga)
   yield spawn(feesSaga)
   yield spawn(localCurrencySaga)
-  yield spawn(syncModeSaga)
+  yield spawn(web3Saga)
 }
