+++ packages/mobile/src/web3/saga.ts
@@ -1,8 +1,5 @@
-import { deriveCEK } from '@celo/utils/src/commentEncryption'
 import { getAccountAddressFromPrivateKey } from '@celo/walletkit'
-import * as Crypto from 'crypto'
 import { generateMnemonic, mnemonicToSeedHex } from 'react-native-bip39'
-import * as RNFS from 'react-native-fs'
 import { REHYDRATE } from 'redux-persist/es/constants'
 import { call, delay, put, race, select, spawn, take, takeLatest } from 'redux-saga/effects'
 import { setAccountCreationTime } from 'src/account/actions'
@@ -25,12 +22,17 @@ import {
   setAccountInWeb3Keystore,
   SetIsZeroSyncAction,
   setLatestBlockNumber,
-  setPrivateCommentKey,
   setZeroSyncMode,
   updateWeb3SyncProgress,
 } from 'src/web3/actions'
 import { addLocalAccount, isZeroSyncMode, web3 } from 'src/web3/contracts'
-import { currentAccountInWeb3KeystoreSelector, currentAccountSelector } from 'src/web3/selectors'
+import { currentAccountSelector } from 'src/web3/selectors'
+import {
+  assignDataKeyFromPrivateKey,
+  ensureAccountInWeb3Keystore,
+  readPrivateKeyFromLocalDisk,
+  savePrivateKeyToLocalDisk,
+} from 'src/web3/zeroSync'
 import { Block } from 'web3/eth/types'
 
 const ETH_PRIVATE_KEY_LENGTH = 64
@@ -107,33 +109,6 @@ export function* waitForWeb3Sync() {
   }
 }
 
-export function* ensureAccountInWeb3Keystore() {
-  const account = yield select(currentAccountSelector)
-  if (account) {
-    const accountInWeb3Keystore = yield select(currentAccountInWeb3KeystoreSelector)
-    if (!accountInWeb3Keystore) {
-      Logger.debug(
-        TAG + '@ensureAccountInWeb3Keystore',
-        'Importing account from private key to web3 keystore'
-      )
-      // Add account to web3 keystore
-      const pincode = yield call(getPincode)
-      const privateKey: string = yield readPrivateKeyFromLocalDisk(account, pincode)
-      // TODO use function addAccountToWeb3Keystore() once it exists
-      // @ts-ignore
-      const web3Acc = yield call(web3.eth.personal.importRawKey, String(privateKey), pincode)
-      yield put(setAccountInWeb3Keystore(web3Acc))
-      yield call(web3.eth.personal.unlockAccount, account, pincode, UNLOCK_DURATION)
-      return web3Acc
-    } else {
-      // TODO check that account and accountInWeb3Keystore are the same
-      return accountInWeb3Keystore
-    }
-  } else {
-    throw new Error('Account not yet initialized') // TODO(anna) decide how to handle this
-  }
-}
-
 export function* getOrCreateAccount() {
   const account = yield select(currentAccountSelector)
   if (account) {
@@ -161,6 +136,25 @@ export function* getOrCreateAccount() {
   }
 }
 
+export function* addAccountToWeb3Keystore(key: string, currentAccount: string, pincode: any) {
+  let account: string
+  try {
+    // @ts-ignore
+    account = yield call(web3.eth.personal.importRawKey, String(key), pincode)
+    yield put(setAccountInWeb3Keystore(account))
+  } catch (e) {
+    if (e.toString().includes('account already exists')) {
+      account = currentAccount
+      Logger.debug(TAG + '@assignAccountFromPrivateKey', 'Importing same account as current one')
+    } else {
+      Logger.error(TAG + '@assignAccountFromPrivateKey', 'Error importing raw key')
+      throw e
+    }
+  }
+  yield call(web3.eth.personal.unlockAccount, account, pincode, UNLOCK_DURATION)
+  return account
+}
+
 export function* assignAccountFromPrivateKey(key: string) {
   const currentAccount = yield select(currentAccountSelector)
 
@@ -175,46 +169,29 @@ export function* assignAccountFromPrivateKey(key: string) {
 
     // Save the account to a local file on the disk.
     // This is done for all sync modes, to allow users to switch in to zero sync mode.
-    // If geth is running it has its own mechanism to save the encrypted key in its keystore.
+    // Note that if geth is running it saves the encrypted key in the web3 keystore.
     account = getAccountAddressFromPrivateKey(key)
     yield savePrivateKeyToLocalDisk(account, key, pincode)
 
     if (isZeroSyncMode()) {
+      // For zero sync mode, add local account
       Logger.debug(TAG + '@assignAccountFromPrivateKey', 'Init web3 with private key')
       addLocalAccount(web3, key)
     } else {
-      try {
-        // TODO make this a function addAccountToWeb3Keystore()
-        // @ts-ignore
-        account = yield call(web3.eth.personal.importRawKey, String(key), pincode)
-        yield put(setAccountInWeb3Keystore(account))
-      } catch (e) {
-        if (e.toString().includes('account already exists')) {
-          account = currentAccount
-          Logger.debug(
-            TAG + '@assignAccountFromPrivateKey',
-            'Importing same account as current one'
-          )
-        } else {
-          Logger.error(TAG + '@assignAccountFromPrivateKey', 'Error importing raw key')
-          throw e
-        }
-      }
-      yield call(web3.eth.personal.unlockAccount, account, pincode, UNLOCK_DURATION)
+      // If geth is running, add to web3 accounts
+      account = yield call(addAccountToWeb3Keystore, key, currentAccount, pincode)
+      web3.eth.defaultAccount = account
     }
 
     Logger.debug(
       TAG + '@assignAccountFromPrivateKey',
       `Created account from mnemonic and added to wallet: ${account}`
     )
-
     yield put(setAccount(account))
     yield put(setAccountCreationTime())
+
     yield call(assignDataKeyFromPrivateKey, key)
 
-    if (!isZeroSyncMode()) {
-      web3.eth.defaultAccount = account
-    }
     return account
   } catch (e) {
     Logger.error(
@@ -223,78 +200,6 @@ export function* assignAccountFromPrivateKey(key: string) {
       e
     )
     return null
-  }
-}
-
-function* assignDataKeyFromPrivateKey(key: string) {
-  const privateCEK = deriveCEK(key).toString('hex')
-  yield put(setPrivateCommentKey(privateCEK))
-}
-
-function getPrivateKeyFilePath(account: string): string {
-  return `${RNFS.DocumentDirectoryPath}/private_key_for_${account}.txt`
-}
-
-function ensureAddressAndKeyMatch(address: string, privateKey: string) {
-  const generatedAddress = getAccountAddressFromPrivateKey(privateKey)
-  if (!generatedAddress) {
-    throw new Error(`Failed to generate address from private key`)
-  }
-  if (address.toLowerCase() !== generatedAddress.toLowerCase()) {
-    throw new Error(
-      `Address from private key: ${generatedAddress}, ` + `address of sender ${address}`
-    )
-  }
-  Logger.debug(TAG + '@ensureAddressAndKeyMatch', `Sender and private key match`)
-}
-
-async function savePrivateKeyToLocalDisk(
-  account: string,
-  privateKey: string,
-  encryptionPassword: string
-) {
-  ensureAddressAndKeyMatch(account, privateKey)
-  const filePath = getPrivateKeyFilePath(account)
-  const plainTextData = privateKey
-  const encryptedData: Buffer = getEncryptedData(plainTextData, encryptionPassword)
-  Logger.debug('savePrivateKeyToLocalDisk', `Writing encrypted private key to ${filePath}`)
-  await RNFS.writeFile(getPrivateKeyFilePath(account), encryptedData.toString('hex'))
-}
-
-// Reads and returns unencrypted private key
-export async function readPrivateKeyFromLocalDisk(
-  account: string,
-  encryptionPassword: string
-): Promise<string> {
-  const filePath = getPrivateKeyFilePath(account)
-  Logger.debug('readPrivateKeyFromLocalDisk', `Reading private key from ${filePath}`)
-  const hexEncodedEncryptedData: string = await RNFS.readFile(filePath)
-  const encryptedDataBuffer: Buffer = new Buffer(hexEncodedEncryptedData, 'hex')
-  const privateKey: string = getDecryptedData(encryptedDataBuffer, encryptionPassword)
-  ensureAddressAndKeyMatch(account, privateKey)
-  return privateKey
-}
-
-// Exported for testing
-export function getEncryptedData(plainTextData: string, password: string): Buffer {
-  try {
-    const cipher = Crypto.createCipher('aes-256-cbc', password)
-    return Buffer.concat([cipher.update(new Buffer(plainTextData, 'utf8')), cipher.final()])
-  } catch (e) {
-    Logger.error(TAG + '@getEncryptedData', 'Failed to write private key', e)
-    throw e // Re-throw
-  }
-}
-
-// Exported for testing
-export function getDecryptedData(encryptedData: Buffer, password: string): string {
-  try {
-    const decipher = Crypto.createDecipher('aes-256-cbc', password)
-    const decrypted = Buffer.concat([decipher.update(encryptedData), decipher.final()])
-    return decrypted.toString('utf8')
-  } catch (e) {
-    Logger.error(TAG + '@getDecryptedData', 'Failed to read private key', e)
-    throw e // Re-throw
   }
 }
 
@@ -379,12 +284,11 @@ export function* getConnectedUnlockedAccount() {
 }
 
 export function* switchToGethFromZeroSync() {
-  Logger.debug(TAG + 'Switching to geth from zero sync..')
   const account = yield call(ensureAccountInWeb3Keystore)
   setZeroSyncMode(false)
   Logger.debug(
     TAG + '@switchToGethFromZeroSync',
-    'Importing account from private key to web3 keystore',
+    'Imported account from private key to web3 keystore',
     account
   )
 }
@@ -408,6 +312,6 @@ export function* watchZeroSyncMode() {
   yield takeLatest(Actions.SET_IS_ZERO_SYNC, switchZeroSyncMode)
 }
 
-export function* syncModeSaga() {
+export function* web3Saga() {
   yield spawn(watchZeroSyncMode)
 }
